Move Buffer.test and test utils to common
